refactor(owned): make minor decomposition
